moved errorHandler to sep script moved PATCH requests for comments and posts to their respective routers updated view controllers with new PATCH path
