+++ controllers/api/commentRoutes.js
@@ -2,28 +2,7 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 const withAuth = require('../../utils/auth');
 const { log, info, warn, error, highlight } = require('@frenzie24/logger')
-
-//create a new comment
-router.post('/', withAuth, async (req, res) => {
-    info('attempting new comment')
-    log(withAuth)
-    log(req.body)
-    try {
-        // we try to create a comment data passed from the client (req.body)
-        const newComment = await Comment.create({
-            ...req.body,
-            user_id: req.session.user_id,
-        });
-        log('Comment created.', 'green')
-        highlight(newComment)
-        if(res.redirected) res.redirect();
-        res.status(200).send({...req.body, user_id: req.session.user_id});
-    } catch (err) {
-        warn('We ran into an error:')
-error(err);
-        res.status(400).json(err);
-    }
-});
+const handleError = require('../../utils/handleError')
 
 // delete by id: when the path is /:id, then the id would go /1 to get the comment with an id of 1
 router.delete('/:id', withAuth, async (req, res) => {
@@ -63,5 +42,57 @@ error(err);
         res.status(500).json(err);
     }
 });
+
+//update comment
+router.patch('/comment', withAuth, async (req, res) => {
+    log('============================');
+    info(`updating comment id: ${req.query.id}`);
+    log(req.body, 'red', 'bgWhite');
+    // log(req.body.content)
+  
+    try {
+  
+      const _id = Math.floor(req.query.id);
+      if (!Number.isInteger(_id)) {
+        warn(`Bad request: id invalid`);
+       return handleError(err, req.session.logged_in, res);
+  
+      }
+      const comment = await Comment.findByPk(_id)
+  
+      comment.content = req.body.content;
+      comment.save();
+      return res.status(200);
+  
+    } catch (err) {
+      return handleError(err, req.session.logged_in, res);
+    }
+  });
+
+//create a new comment
+router.post('/', withAuth, async (req, res) => {
+    info('attempting new comment')
+    log(withAuth)
+    log(req.body)
+    try {
+        // we try to create a comment data passed from the client (req.body)
+        const newComment = await Comment.create({
+            ...req.body,
+            user_id: req.session.user_id,
+        });
+        log('Comment created.', 'green')
+        highlight(newComment)
+        if(res.redirected) res.redirect();
+        res.status(200).send({...req.body, user_id: req.session.user_id});
+    } catch (err) {
+        warn('We ran into an error:')
+error(err);
+        res.status(400).json(err);
+    }
+});
+
+
+
+
 
 module.exports = router;
